requirements.txt
fastapi
uvicorn[standard]
transformers
datasets
torch
accelerate
peft
python-dotenv
openai 